refactor: Move all helpers in /react
To remove specific watch for those files
