+++ docs/styleguide.config.js
@@ -219,7 +219,7 @@ module.exports = {
   usageMode: 'collapse',
   context: {
     utils: path.resolve(__dirname, 'utils'),
-    isTesting: path.resolve(__dirname, '../helpers/isTesting'),
+    isTesting: path.resolve(__dirname, '../react/helpers/isTesting'),
     content: path.resolve(__dirname, 'fixtures/content')
   }
 }
