--- helpers/index.js
@@ -1,1 +1,0 @@
-export { default as withOffline } from './withOffline'
