+++ react/helpers/isTesting.js
